+++ apps/vs-agent/src/main.ts
@@ -2,17 +2,10 @@ import 'reflect-metadata'
 
 import type { ServerConfig } from './utils/ServerConfig'
 
-import { HttpOutboundTransport, KeyDerivationMethod, LogLevel, utils } from '@credo-ts/core'
-import { HttpInboundTransport } from '@credo-ts/node'
-import { DynamicModule, INestApplication, ValidationPipe, VersioningType } from '@nestjs/common'
+import { KeyDerivationMethod } from '@credo-ts/core'
 import { NestFactory } from '@nestjs/core'
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger'
-import express from 'express'
 import * as fs from 'fs'
-import { IncomingMessage } from 'http'
-import * as net from 'net'
 import * as path from 'path'
-import WebSocket from 'ws'
 
 import packageJson from '../package.json'
 
@@ -40,97 +33,17 @@ import {
   USE_CORS,
   USER_PROFILE_AUTODISCLOSE,
 } from './config/constants'
-import { DidCommModule } from './didcomm.module'
 import { connectionEvents } from './events/ConnectionEvents'
 import { messageEvents } from './events/MessageEvents'
 import { vcAuthnEvents } from './events/VCAuthnEvents'
 import { VsAgent } from './utils/VsAgent'
-import { VsAgentWsInboundTransport } from './utils/VsAgentWsInboundTransport'
-import { VsAgentWsOutboundTransport } from './utils/VsAgentWsOutboundTransport'
 import { TsLogger } from './utils/logger'
-import { setupAgent } from './utils/setupAgent'
+import { commonAppConfig, setupAgent } from './utils/setupAgent'
 
-export const enableHttpAndWs = async (
-  app: INestApplication,
-  agent: VsAgent,
-  endpoints: string[],
-  port: number,
-  logLevel?: LogLevel,
-) => {
-  const logger = new TsLogger(logLevel ?? LogLevel.warn, 'Agent')
-
-  const enableHttp = endpoints.find(endpoint => endpoint.startsWith('http'))
-  const enableWs = endpoints.find(endpoint => endpoint.startsWith('ws'))
-
-  let webSocketServer: WebSocket.Server | undefined
-  let httpInboundTransport: HttpInboundTransport | undefined
-  if (enableHttp) {
-    logger.info('Inbound HTTP transport enabled')
-    const expressApp = app.getHttpAdapter().getInstance()
-    httpInboundTransport = new HttpInboundTransport({ app: expressApp, port })
-    agent.registerInboundTransport(httpInboundTransport)
-  }
-
-  if (enableWs) {
-    logger.info('Inbound WebSocket transport enabled')
-    webSocketServer = new WebSocket.Server({ noServer: true })
-    agent.registerInboundTransport(new VsAgentWsInboundTransport({ server: webSocketServer }))
-  }
-
-  agent.registerOutboundTransport(new HttpOutboundTransport())
-  agent.registerOutboundTransport(new VsAgentWsOutboundTransport())
-
-  // await agent.initialize()
-
-  const httpServer = httpInboundTransport ? httpInboundTransport.server : await app.listen(port)
-
-  // Add WebSocket support if required
-  if (enableWs) {
-    httpServer.on('upgrade', (request: IncomingMessage, socket: net.Socket, head: Buffer) => {
-      webSocketServer?.handleUpgrade(request, socket, head, ws => {
-        const socketId = utils.uuid()
-        webSocketServer?.emit('connection', ws, request, socketId)
-      })
-    })
-  }
-}
-
-export const startAdminServer = async (dynamicModule: DynamicModule, serverConfig: ServerConfig) => {
-  const app = await NestFactory.create(dynamicModule)
-
-  // Version
-  app.enableVersioning({
-    type: VersioningType.URI,
-  })
-
-  // Swagger
-  const config = new DocumentBuilder()
-    .setTitle('API Documentation')
-    .setDescription('API Documentation')
-    .setVersion('1.0')
-    .build()
-  const document = SwaggerModule.createDocument(app, config)
-  SwaggerModule.setup('api', app, document)
-
-  // agent didcomm
-  app.use(express.json({ limit: '5mb' }))
-  app.use(express.urlencoded({ extended: true, limit: '5mb' }))
-  app.getHttpAdapter().getInstance().set('json spaces', 2)
-
-  // Dto
-  app.useGlobalPipes(new ValidationPipe())
-
-  // Cors
-  if (serverConfig.cors) {
-    app.enableCors({
-      origin: '*',
-      methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
-      allowedHeaders: 'Content-Type,Authorization',
-    })
-  }
-
+export const startAdminServer = async (agent: VsAgent, serverConfig: ServerConfig) => {
+  const app = await NestFactory.create(VsAgentModule.register(agent))
   // Port expose
-  return app
+  return commonAppConfig(app, serverConfig)
 }
 
 const run = async () => {
@@ -185,16 +98,14 @@ const run = async () => {
   }
 
   // Start admin server
-  const app = await startAdminServer(VsAgentModule.register(agent), conf)
+  const app = await startAdminServer(agent, conf)
   await app.listen(conf.port)
 
   // Start admin didcomm agent server
-  const didCommApp = await startAdminServer(DidCommModule.register(agent), {
-    port: AGENT_PORT,
-    cors: USE_CORS,
-    logger: serverLogger,
-  })
-  enableHttpAndWs(didCommApp, agent, AGENT_ENDPOINTS, AGENT_PORT, AGENT_LOG_LEVEL)
+  // const didCommApp = await startAdminServer(DidCommModule.register(agent), {
+  //   ...conf,
+  //   port: AGENT_PORT,
+  // })
 
   // Listen to events emitted by the agent
   connectionEvents(agent, conf)
